% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO712PACKAGE-package.R
\docType{package}
\name{GEO712PACKAGE-package}
\alias{GEO712PACKAGE}
\alias{GEO712PACKAGE-package}
\title{GEO712PACKAGE: A R package for the GEO712 course at McMAster University}
\description{
In this R package, we share data processed after a multiple linear regression. We used data from the data science platform Kaggle, called "House Sales in King County, USA", to predict the price of houses sold in some US cities using a regression model. The dataset contains house sale prices for King County, which includes Seattle. It includes homes sold between May 2014 and May 2015, with 21,613 records and 21 variables relating to the homes sold.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/GEO712PACKAGE}
  \item \url{https://dias-bruno.github.io/GEO712PACKAGE/}
  \item Report bugs at \url{https://github.com/dias-bruno/GEO712PACKAGE/issues}
}

}
\keyword{internal}
