% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO712PACKAGE-package.R
\docType{package}
\name{GEO712PACKAGE-package}
\alias{GEO712PACKAGE}
\alias{GEO712PACKAGE-package}
\title{GEO712PACKAGE: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/GEO712PACKAGE}
  \item \url{https://dias-bruno.github.io/GEO712PACKAGE/}
  \item Report bugs at \url{https://github.com/dias-bruno/GEO712PACKAGE/issues}
}

}
\keyword{internal}
